Hangman
This is my first project with Python. I spent a couple of hours learning from https://www.learnpython.org/ and playing around on the PyCharm IDE. Any feedback would be appreciated, and feel free to use this project for yourself. To add more words, you can simply add to the words list in line 3. Just be sure to update the random integer range in line 10 to go up to the new total number of words minus one.